crewai-orchestrator/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
crewai==0.28.8
python-multipart==0.0.6
PyYAML==6.0.2
jsonschema==4.22.0
